+++ src/widgets/settings/generalsettingswidget.cpp
@@ -150,6 +150,9 @@ void GeneralSettingsWidget::readSettings() {
         Utils::Misc::doAutomaticNoteFolderDatabaseClosing());
     ui->legacyLinkingCheckBox->setChecked(settings.value(QStringLiteral("legacyLinking")).toBool());
 
+    ui->stripLeadingEmojiFromNoteFilenameCheckBox->setChecked(
+        settings.value(QStringLiteral("stripLeadingEmojiFromNoteFilename"), true).toBool());
+
     const bool enableReadOnlyMode =
         settings.value(QStringLiteral("enableReadOnlyMode"), true).toBool();
     ui->enableReadOnlyModeCheckBox->setChecked(enableReadOnlyMode);
@@ -239,6 +242,8 @@ void GeneralSettingsWidget::storeSettings() {
     settings.setValue(QStringLiteral("automaticNoteFolderDatabaseClosing"),
                       ui->automaticNoteFolderDatabaseClosingCheckBox->isChecked());
     settings.setValue(QStringLiteral("legacyLinking"), ui->legacyLinkingCheckBox->isChecked());
+    settings.setValue(QStringLiteral("stripLeadingEmojiFromNoteFilename"),
+                      ui->stripLeadingEmojiFromNoteFilenameCheckBox->isChecked());
     settings.setValue(QStringLiteral("enableReadOnlyMode"),
                       ui->enableReadOnlyModeCheckBox->isChecked());
     settings.setValue(QStringLiteral("startInReadOnlyMode"),